Model struct

+++ CrossingHighway/systemclass.cpp
@@ -216,20 +216,20 @@ bool SystemClass::Frame()
 		return false;
 	}
 
-	if (m_Input->IsUpPressed() && (m_Graphics->m_SystemPlayerV.z - m_Graphics->m_PlayerV.z) <= 0.1) {
+	if (m_Input->IsUpPressed() && ((float)m_Graphics->m_SystemPlayerV.z - m_Graphics->m_PlayerV.z) <= 0.1f) {
 		m_Graphics->m_SystemPlayerV.z += 5.0f;
 		m_Graphics->m_PlayerRotation.y = 0.0f * D3DX_PI / 180;
 	}
-	if (m_Input->IsDownPressed() && (m_Graphics->m_PlayerV.z - m_Graphics->m_SystemPlayerV.z) <= 0.1) {
+	if (m_Input->IsDownPressed() && ((float)m_Graphics->m_PlayerV.z - m_Graphics->m_SystemPlayerV.z) <= 0.1f) {
 		m_Graphics->m_SystemPlayerV.z -= 5.0f;
 		m_Graphics->m_PlayerRotation.y = 180.0f * D3DX_PI / 180;
 	}
-	if (m_Input->IsLeftPressed() && (m_Graphics->m_PlayerV.x - m_Graphics->m_SystemPlayerV.x) <= 0.1) {
+	if (m_Input->IsLeftPressed() && ((float)m_Graphics->m_PlayerV.x - m_Graphics->m_SystemPlayerV.x) <= 0.1f) {
 		m_Graphics->m_SystemPlayerV.x -= 5.0f;
 		m_Graphics->m_PlayerRotation.y = 270.0f * D3DX_PI / 180;
 
 	}
-	if (m_Input->IsRightPressed() && (m_Graphics->m_SystemPlayerV.x - m_Graphics->m_PlayerV.x) <= 0.1) {
+	if (m_Input->IsRightPressed() && ((float)m_Graphics->m_SystemPlayerV.x - m_Graphics->m_PlayerV.x) <= 0.1f) {
 		m_Graphics->m_SystemPlayerV.x += 5.0f;
 		m_Graphics->m_PlayerRotation.y = 90.0f * D3DX_PI / 180;
 	}
